feat: roundId increment

+++ src/interfaces/OracleMiddleware/IOracleMiddleware.sol
@@ -37,8 +37,8 @@ interface IOracleMiddleware is IBaseOracleMiddleware, IOracleMiddlewareErrors, I
      */
     function getConfRatioBps() external view returns (uint16);
 
-    /// @notice Return the delay during which a low latency oracle validation is available
-    function getLowLatencyDelay() external pure returns (uint256);
+    /// @notice Return the delay during which a low latency oracle price validation is available
+    function getLowLatencyDelay() external pure returns (uint16);
 
     /* -------------------------------------------------------------------------- */
     /*                               Owner features                               */
